Update MRI upstreams for Phosphorus
Adopt latest versions, namely;
- odlparent-9.0.2
- infrautils-2.0.2
- mdsal-8.0.0
- controller-4.0.0
- aaa-0.14.0
- netconf-2.0.0

Also introduce model-{inventory,topology}, which were removed from
upstream.

Since these versions also force the Uint8/Uint32 vs. Short/Long
equivalence issue, we mass-migrate all APIs to use Uint constructs.
This in turn requires minor adjustments to mocks, as we no longer
get automatic smart zeroes.

Another thing that is cropping up here is that augment-identifier
is now reflected in package names properly, requiring a proper update
of imports and references.

+++ extension/openflowplugin-extension-api/src/main/java/org/opendaylight/openflowplugin/extension/api/ConverterMessageToOFJava.java
@@ -11,6 +11,7 @@ import org.opendaylight.openflowplugin.extension.api.exception.ConversionExcepti
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflow.common.types.rev130731.ExperimenterId;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.openflowplugin.experimenter.types.rev151020.experimenter.core.message.ExperimenterMessageOfChoice;
 import org.opendaylight.yangtools.yang.binding.DataContainer;
+import org.opendaylight.yangtools.yang.common.Uint32;
 
 /**
  * Convert message from MD-SAL model into OFJava-API model.
@@ -41,5 +42,5 @@ public interface ConverterMessageToOFJava<F extends ExperimenterMessageOfChoice,
      * Returns the corresponding experimenter message type.
      * @return experimenter message type.
      */
-    long getType();
+    Uint32 getType();
 }
